feat(compiler): 添加位向量实现和基准测试功能
- 实现 BitVector 类及其完整操作接口
- 添加 BitVector 单元测试验证功能正确性
- 创建 ADT、ISel 和 RA 模块的性能基准测试
- 实现 X86 指令编码器和相关指令表
- 添加机器函数创建和虚拟寄存器管理功能
- 优化分配器对齐计算逻辑
- 集成 clang-tidy 代码静态检查脚本

+++ lib/ADT/BitVector.cpp
@@ -94,7 +94,7 @@ unsigned BitVector::size() const noexcept { return numBits_; }
 unsigned BitVector::count() const {
     unsigned c = 0;
     for (unsigned i = 0; i < numWords_; ++i)
-        c += __builtin_popcountll(words_[i]);
+        c += static_cast<unsigned>(__builtin_popcountll(words_[i]));
     return c;
 }
 
@@ -153,7 +153,7 @@ void BitVector::flip() {
 int BitVector::find_first() const {
     for (unsigned i = 0; i < numWords_; ++i) {
         if (words_[i]) {
-            const int bit = __builtin_ctzll(words_[i]);
+            const unsigned bit = static_cast<unsigned>(__builtin_ctzll(words_[i]));
             const int result = static_cast<int>(i * BITS_PER_WORD + bit);
             return result < static_cast<int>(numBits_) ? result : -1;
         }
@@ -167,7 +167,7 @@ int BitVector::find_next(unsigned idx) const {
     word_type word = words_[wordIdx] & (~static_cast<word_type>(0) << (idx % BITS_PER_WORD));
     while (true) {
         if (word) {
-            const int bit = __builtin_ctzll(word);
+            const unsigned bit = static_cast<unsigned>(__builtin_ctzll(word));
             const int result = static_cast<int>(wordIdx * BITS_PER_WORD + bit);
             return result < static_cast<int>(numBits_) ? result : -1;
         }
